basic/sample_tfrecord/requirements.txt
tensorflow==2.12.0
numpy>=1.19.2
pyyaml==6.0.2
pillow==10.4.0